pyproject.toml
[project]
name = "griffonner"
dynamic = ["version"]
description = "A templating tool for Griffe."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Will Langdale", email = "willdon@gmail.com" }
]
requires-python = ">=3.9,<3.14"
dependencies = [
    "griffe>=1.12.1",
    "typer>=0.16.1",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "watchdog>=3.0.0",
]

[project.scripts]
griffonner = "griffonner:main"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

# Presence enables setuptools-scm
[tool.setuptools_scm]


[tool.uv]
package = true

[tool.ruff]
exclude = [
    "*.ipynb"
]
line-length = 88
indent-width = 4
target-version = "py39"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "D"
]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/{test,docs}/*" = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.2",
    "mkdocs-material>=9.6.17",
    "mkdocstrings[python]>=0.30.0",
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.0.0",
    "types-jinja2>=2.11.9",
]
